# RFC: Add a `logrange` function