--- a/base/exports.jl
+++ b/base/exports.jl
@@ -58,7 +58,6 @@ export
     IOBuffer,
     IOStream,
     LinRange,
-    LogRange,
     Irrational,
     LazyString,
     Matrix,