--- a/base/range.jl
+++ b/base/range.jl
@@ -1489,48 +1489,150 @@ julia> mod(3, 0:2)  # mod(3, 3)
 mod(i::Integer, r::OneTo) = mod1(i, last(r))
 mod(i::Integer, r::AbstractUnitRange{<:Integer}) = mod(i-first(r), length(r)) + first(r)
 
+
 """
-    logrange(start => stop, length)
+    logrange(start, stop; length, ratio)
 
-Returns an iterator which runs from `start` to `stop` with `length` elements,
-spaced logarithmically rather than linearly as for [`range`](@ref).
-That is, the ratio of successive elements is constant, not the difference.
+Create an iterator whose elements are spaced logarithmically, rather than linearly
+as for [`range`](@ref), i.e. the ratio of successive elements is constant, not the difference.
 
-!!! compat "Julia 1.3"
-     This function requires at least Julia 1.7.
+One of two keywords must be provided:
+
+* Given the `length`, the necessary ratio is computed accurately to reach the final and
+  intermediate values rationally. To avoid this overhead, see the [`LogRange`](@ref) constructor.
+
+* Otherwise, given the `ratio`, the final value is the last one `<= stop`.
+
+!!! compat "Julia 1.7"
+    This function requires at least Julia 1.7.
+
+This is similar to `geomspace` in Python, and to `PowerRange` in Mathematica.
 
 # Examples
-```
-julia> foreach(println, logrange(2 => 16, 4))
-2.0
-4.0
-8.0
-16.0
+```jldoctest
+julia> for x in logrange(2, 19; ratio=2)
+         println(x)
+       end
+2
+4
+8
+16
 
-julia> collect(logrange(1000 => 1, 4))
-4-element Vector{Float64}:
-    1.0
-    9.999999999999998
-   99.99999999999997
+julia> collect(logrange(1000, 1; length=7))
+7-element Vector{Float64}:
  1000.0
+  316.22776601683796
+  100.0
+   31.622776601683793
+   10.0
+    3.1622776601683795
+    1.0
 
-julia> ans ≈ 10 .^ (3:-1:0)
+julia> ans ≈ 10 .^ (3:-0.5:0)
 true
 
-julia> collect(logrange(-1 => -2f0, 3))
+julia> collect(logrange(-1, -2f0, length=3))
 3-element Vector{Float32}:
  -1.0
  -1.4142135
  -2.0
+
+julia> map(rad2deg∘angle, logrange(1, -1+0im, length=5))
+5-element Vector{Float64}:
+   0.0
+  45.0
+  90.0
+ 135.0
+ 180.0
 ```
 """
-function logrange(p::Pair{<:Real, <:Real}, n::Integer)
-    lo, hi = promote(p.first, p.second)
-    if lo>0 && hi>0
-        (exp(x) for x in range(log(lo), log(hi), length=n))
-    elseif lo<0 && hi<0
-        (-exp(x) for x in range(log(-lo), log(-hi), length=n))
-    else
-        throw(DomainError(p, "logrange requires that first and last elements are both positive, or both negative"))
+function logrange(start::Number, stop::Number;
+    length::Union{Nothing,Integer}=nothing, ratio::Union{Nothing,Number}=nothing)
+    if length !== nothing
+        length >= 2 || throw(ArgumentError("logrange must have length of at least 2"))
+        _start, _stop = map(float, promote(start, stop))
+        _ratio = ratio_nth_root(stop, start, Int(length)-1)
+        if ratio !== nothing
+            check = oftype(float(ratio), _ratio)
+            ratio == check || throw(ArgumentError("ratio = $ratio does not match $check computed from length = $length"))
+        end
+        LogRange(_start, _ratio, Int(length))
+    elseif ratio !== nothing
+        _start, _ratio = promote(start, ratio)
+        num = trunc(Int, log(ratio, stop/_start))
+        fin = prod(_ratio for _ in 1:num; init=start)
+        _length = fin <= stop ? num+1 : num
+        LogRange(_start, _ratio, _length)
     end
 end
+
+ratio_nth_root(a::Number, b::Number, n::Int) = (a/b)^(1/n)
+ratio_nth_root(a::Float32, b::Float32, n::Int) = (Float64(a)/Float64(b))^(1/n)
+function ratio_nth_root(a::Real, b::Real, m::Int)
+    over = TwicePrecision(a) / TwicePrecision(b)
+    r1 = TwicePrecision((over.hi)^(1/m))
+    r1pow = prod(r1 for _ in 1:m-1)
+    r2 = (m-1)*r1/m + over / (m * r1pow)
+end
+
+"""
+    LogRange{T,S}(start::T, ratio::S, len::Int)
+
+Iterator which multiplies by `ratio` at each step, giving in total `len` objects.
+Constructing this directly instead of using [`logrange`](@ref) avoids the cost of
+computing `ratio` to high precision.
+
+Note that its output is of type `T`, so iterating for intance `LogRange(2, pi, 10)` will
+lead to an `InexactError`, since `2*pi` cannot be converted to `Int`.
+
+# Examples
+
+```jldoctest
+julia> collect(LogRange(0.1, sqrt(10), 5))  # no corrections for rounding errors
+5-element Vector{Float64}:
+  0.1
+  0.316227766016838
+  1.0000000000000002
+  3.1622776601683804
+ 10.000000000000004
+
+julia> r = logrange(0.1, 10, length=5)  # type T != type S
+LogRange{Float64, Base.TwicePrecision{Float64}}(0.1, Base.TwicePrecision{Float64}(3.162277660168379, 2.0941562178568784e-16), 5)
+
+julia> collect(r)
+5-element Vector{Float64}:
+  0.1
+  0.31622776601683794
+  1.0
+  3.1622776601683795
+ 10.0
+```
+"""
+struct LogRange{T,S}
+    start::T
+    ratio::S
+    len::Int
+end
+
+length(r::LogRange) = r.len
+size(r::LogRange) = (r.len,)
+
+function iterate(r::LogRange{T,S}) where {T,S}
+    y = convert(promote_type(T,S), r.start)
+    r.start, (y,1)
+end
+
+function iterate(r::LogRange{T}, (x,n)) where {T}
+    if n >= r.len
+        nothing
+    else
+        y = x * r.ratio
+        convert(T,y), (y,n+1)
+    end
+end
+
+eltype(::Type{<:LogRange{T}}) where {T} = T
+eltype(r::LogRange{T}) where {T} = T
+
+ndims(::Type{<:LogRange}) = 1
+ndims(r::LogRange) = 1