I think it would be useful to have an easy way to make a range of points spaced logarithmically. So, for discussion, this adds a function which does this:
```julia
julia> collect(logrange(1 => 100, 5))
5-element Vector{Float64}:
   1.0
   3.1622776601683795
  10.000000000000002
  31.622776601683803
 100.00000000000004
```
Before Julia 1.0, there was a `logspace` function which something different, and less useful:
```julia
julia> logspace(1, 4, 4, base=2)
4-element Array{Float64,1}:
  2.0
  4.0
  8.0
 16.0
```
This was removed in #25896 in favour of just writing `2 .^ (1:4)`. If you know what powers you want, that is fine. But if you know what start and endpoints you want, then it's more of a hassle to write something like `2 .^ range(log2(2), log2(16), length=4)`. Well, for powers of 2, I think we can all do that in our heads, but if you want 20 plot points between about 3.5 and 1700, that's when this would be neat. 

After deciding whether or not this is a good idea at all:

* It could be its own `LinRange` or something, instead of just a generator. Or it could eagerly make a `Vector`. 
* I've given this the same notation as #39069 but it could equally well be `logrange(lo, hi; length)` instead, or in addition.
* Matlab's `logspace` behaves like the old one, so perhaps best to avoid that name. Mathematica's `PowerRange[a,b,r]` takes start, stop, and ratio, that's another possible name. It could like `range` be made to accept keywords `ratio=1.1` instead of `length=10`.
* Ideally it would use something like TwicePrecision to hit the endpoints more precisely?