@antoine-levitt 2021-01-02
I agree this is useful. If we do this we should go with `linrange` (instead of `range(start, stop, length)`) for consistency. That might be the best solution here. Added bonus is that we can easily add features to these functions if we need (eg I'd love an `endpoints=:none/:left/:right/:both`), which is much harder to do with plain old `range`.