@mkitti 2021-01-02
There probably should be a `LogRange` object to help with this. As currently implemented, there is some numerical issues for some common bases.

```julia
julia> collect( logrange(1 => 1000, 4) )
4-element Array{Float64,1}:
   1.0
   9.999999999999998
  99.99999999999996
 999.9999999999998
```

May I suggest the following:

```julia
       function logrange(p::Pair{<:Real, <:Real}, n::Integer)
           lo, hi = promote(p.first, p.second)
           ratio = hi/lo
           if ratio % 10 == 0
               exp = p->10^p
               log = log10
           elseif ratio % 2 == 0
               exp = p->2^p
               log = log2
           end
           if lo>0 && hi>0 || ( lo<0 && hi<0 )
               (lo*exp(x) for x in range(0, log(hi/lo), length=n))
           else
               throw(DomainError(p, "logrange requires that first and last elements are both positive, or both negative"))
           end
       end
```

This increases numerical stability for `base = 2` and `base = 10`

```julia
julia> collect( logrange(1 => 1000, 4) )
4-element Array{Float64,1}:
    1.0
   10.0
  100.0
 1000.0

julia> collect( logrange(1 => 1024, 3) )
3-element Array{Float64,1}:
    1.0
   32.0
 1024.0

julia> collect( logrange(1 => 1024, 6) )
6-element Array{Float64,1}:
    1.0
    4.0
   16.0
   64.0
  256.0
 1024.0
```