@mcabbott 2021-01-02
Yes, the implementation here is just a sketch so far. Ideally it would always hit both endpoints to full accuracy, although I'm not sure how hard that would be.