@mkitti 2021-01-02
> Yes, the implementation here is just a sketch so far. Ideally it would always hit both endpoints to full accuracy, although I'm not sure how hard that would be.

The trick might be figuring out which endpoint we are closer to and then deciding whether we should count up or down. For start, the calculation should be `start*exp(0)`. For stop, the calculation should be likewise `stop*exp(0)`. We cannot rely on `exp(log(x)) === x`. However, `exp(0.0)` is reliably `1.0`.

The need for selecting the correct base and for making these kind of decisions is what ultimately justifies having this method, I think.