@mkitti 2021-01-02
> Ideally it would always hit both endpoints to full accuracy

I updated my suggested code above. I think a ternary operator takes care of the endpoints. 
```julia
               log_ratio = log(ratio)
               (x < log_ratio/2 ? lo*exp(x) : hi*exp(x-log_ratio) for x in range(0, log_ratio, length=n))
```