@mkitti 2021-01-02
Hmm... there is still some asymmetry to this.

```julia
julia> collect(logrange(1 => 1000, 4))
4-element Array{Float64,1}:
    1.0
   10.0
  100.0
 1000.0

julia> collect(logrange(1000 => 1, 4))
4-element Array{Float64,1}:
 999.9999999999998
 100.00000000000004
  10.000000000000005
   1.0000000000000002
```