@mkitti 2021-01-02
I fixed it by always taking the larger ratio.

```julia
                  invert = abs(lo) > abs(hi)
                  ratio = invert ? lo/hi : hi/lo
```

So now everything seems to work well:
```julia
julia> collect( logrange( 1000 => 1, 4) )
4-element Array{Float64,1}:
 1000.0
  100.0
   10.0
    1.0

julia> collect( logrange( 1 => 1000, 4) )
4-element Array{Float64,1}:
    1.0
   10.0
  100.0
 1000.0

julia> collect( logrange( -1 => -1000, 4) )
4-element Array{Float64,1}:
    -1.0
   -10.0
  -100.0
 -1000.0

julia> collect( logrange( -1000 => -1, 4) )
4-element Array{Float64,1}:
 -1000.0
  -100.0
   -10.0
    -1.0
```