@mkitti 2021-01-02
The Numpy equivalent of this function is called `numpy.geomspace` https://numpy.org/doc/stable/reference/generated/numpy.geomspace.html