@darsnack 2021-01-02
Can I suggest that 0 be allowed as an endpoint? I've always found it frustrating when I write code sweeping a parameter and `xs = linrange(0 => 1, 10)` works but `logrange(0 => 1, 10)` errors (here I'm just using the proposed names for clarity). Programmatically, this isn't surprising, but if in English I say "give me 4 points between 0 and 1 logarithmically spaced apart" then I expect `[0, 0.01, 0.1, 1.0]` to be the result. It would be nice if I didn't have to write my own special handling to achieve this.