@antoine-levitt 2021-01-02
@darsnack what you are suggesting requires an explicit base, which logrange doesn't because it's base-independent (the notion of "logarithmically equispaced" does not depend on which logarithm you choose)