@mcabbott 2021-01-02
If this takes keywords like range does, then something like `logspace(0 => 1, length=4, ratio=10) == [0.001, 0.01, 0.1, 1.0]` might not be too odd. You need to provide three of start/stop/length/ratio, and `0` is acting as some kind of nothing? The same could be true of `logspace(1 => Inf, length=4, ratio=10) == [1, 10, 100, 1000]`.

Including zero in the results `[0, 0.01, 0.1, 1.0]` would be much more surprising, as the ratio is infinite for that step.