@darsnack 2021-01-02
Yeah I don't necessarily want to complicate the proposal with too many API options now that I understand the rationale.

I guess it depends on which arguments are chosen for specification. To me, whenever I specify `start` and `stop`, I expect them to be in the range no matter what. So even though the ratio is infinite, the inclusion of 0.0 is less surprising than a changing lower bound depending on the number of steps.