@mkitti 2021-01-04
The latest changes look very nice. I see that the iteration is based on multiplying the ratio by the prior element. How confident are you that you can avoid accumulated errors through long iterations?

The iteration for `UnitRange` and `StepRangeLen` uses multiplication rather than addition:
https://github.com/JuliaLang/julia/blob/master/base/range.jl#L664

`StepRangeLen`:
https://github.com/JuliaLang/julia/blob/master/base/range.jl#L720-L723

`LinRange`:
https://github.com/JuliaLang/julia/blob/master/base/range.jl#L730-L738

Only for `OrdinalRange` where the step is an `Integer` is the iteration done by addition:
https://github.com/JuliaLang/julia/blob/master/base/range.jl#L669-L674

In this case, might there be situations where multiplication or exponentiation might result in better stability?