@mcabbott 2021-01-04
All else equal it would be nicer to have random access, which is what multiplication buys you in linear ranges. But it's tricky to make that fast & accurate here, I think the answer may simply be to collect this iterator if you need that.

Notice that it iterates in twice the precision of the output, so there are quite a few spare bits for the inevitable errors. I haven't been very systematic about stress-testing it, I don't promise it's `< eps` everywhere. It's noticeably less accurate for complex ranges, and possibly that's a mistake like missing a `conj` somewhere; the fact that it accepts complex numbers at all seems like a generalisation which happens for free, not obviously a desired feature.

This gist compares some different approaches: https://gist.github.com/mcabbott/a40c784b4fd27656b581ad0dac09aae5 . (And, I will have to look into why this doesn't build, a bit confused about imports to `Iterators` module but can surely sort it out.)