@JeffBezanson 2021-01-04
I think we should not use Pair for this; I don't believe any other similar functions work like that?