@mcabbott 2021-01-04
Pair is intended to match #39069 for linear `range`, if that gets merged.

This could instead be `logrange(a, b; length=10)`. It does seem a little odd to need a keyword if there is just one option, unlike `range`. It could of course also have methods like `logrange(a, b; ratio=2)` and `logrange(a; length=10, ratio=2)`, the latter is currently spelled `LogRange(a, 2, 10)`.