@mcabbott 2024-01-18
Before triage:

If the main objection is that this is too much code, then perhaps we should discuss variants. The minimal standard is that `LogRange <: AbstractVector` should be lazy, and hit given endpoints exactly.

1. `LogRange{T}` only, with no TwicePrecision at all -- closest to `LinRange`, least code. 
2. `LogRange{T,X}` with TwicePrecision, export only one symbol.
3. Export only `logrange(start, stop; length)`, and keep the struct `LogRange{T,X}` internal -- minor change from this PR.
4. `logrange(start, stop; length)` and `LogRange{T,X}` with TwicePrecision -- the present state of this PR.
5. `logrange(start, stop; length)` with TwicePrecision, and `LogRange(start, stop, length)` which opts out of TwicePrecision -- closer analogy to `range` vs `LinRange`.

Note again that in all of these, TwicePrecision is basically cosmetic -- it means you more often hit pretty intermediate points, when such exist. There is never any magic like `0:0.1:1` understanding that you mean `1//10`, since `start, stop` are always given.

For options 1 & 5, here are some quick benchmarks on how expensive TwicePrecision is.

<details>

```julia
julia> @btime sum(logrange(1, 2, 3));  # with TwicePrecision -- mostly construction cost
  min 16.408 ns, mean 16.739 ns (0 allocations)

julia> @btime sum(logrange(1, 2, 10^4));  # with TwicePrecision -- mostly indexing cost, 9ns/item
  min 89.833 μs, mean 90.348 μs (0 allocations)

julia> @btime sum(LogRange(1, 2, 3));  # without TwicePrecision (option 5, not present PR)
  min 4.916 ns, mean 5.038 ns (0 allocations)

julia> @btime sum(LogRange(1, 2, 10^4));  # without TwicePrecision, 3.5ns/item
  min 35.041 μs, mean 35.260 μs (0 allocations)
```

For comparison, linear ranges:

```julia
julia> @btime sum(0:0.1:1);  # magic inference of length
  min 73.742 ns, mean 85.712 ns (0 allocations)

julia> @btime sum(range(0, 1, length=11));  # constructs a StepRangeLen, but no magic
  min 1.166 ns, mean 1.271 ns (0 allocations)

julia> @btime sum(LinRange(0, 1, 11));
  min 1.167 ns, mean 1.271 ns (0 allocations)

julia> @btime copyto!($(rand(10^4)), $(range(1, 2, 10^4)));  # indexing costs 1.1ns/item
  min 11.125 μs, mean 11.243 μs (0 allocations)

julia> @btime copyto!($(rand(10^4)), $(LinRange(1, 2, 10^4)));  # indexing costs 0.7ns/item
  min 7.406 μs, mean 7.492 μs (0 allocations)

```

</details>

Making a standard library just for `logrange` seems like a lot of hassle for a small thing. I don't know how practical it would be to move e.g. StepRangeLen and TwicePrecision out to a standard library, but that would be a much larger upgrade.

After triage:

It sounds like option 3 is preferred.

Making a larger standard library for all ranges (or at least floating point ones) is probably above my pay grade, but sounds like it may be a longer-term goal? Xref recent discussion on Random https://github.com/JuliaLang/julia/pull/51432 and LinearAlgebra https://github.com/JuliaLang/julia/pull/51904 ... if those can be excised, then perhaps some new FloatingPointRanges can be too. This code would move along with `StepRangeLen` and friends.